Reorganize components into separate directories for better structure
